service2/requirements.txt
Flask==3.0.0
Flask-CORS==3.0.10
Werkzeug==3.0.0
mysql-connector-python